Add files via upload
Handwritten digit prediction involves using machine learning techniques, particularly classification algorithms, to recognize and identify digits that are handwritten. This task is often treated as a supervised learning problem, where the model is trained on a labeled dataset of handwritten digits and then used to predict the digit class of new, unseen samples.